mv interface to shared file and add maybe type back
